requirements.txt
attrs>=21.3.0
black>=22.10.0
httpx>=0.15.4
jsonschema==4.16.0
openapi-python-client>=0.11.6
pylint>=2.15.10
pylint-exit>=1.2.0
pytest>=7.2.0
pytest-cov==4.0.0
pytest-xdist==3.2.1
